Agrego listeners que modifican el balance cuando se agrega una compra. Finalizo el import de excel a psql. Acomodo models para que tengan los campos requeridos

+++ database/migrations/2020_08_12_215802_create_customers_table.php
@@ -18,7 +18,7 @@ class CreateCustomersTable extends Migration
             $table->timestamps();
             $table->string('name');
             $table->foreignId('address_id')->nullable();
-            $table->float('balance')->default(0);
+            $table->integer('balance')->default(0);
             $table->integer('twenty_balance')->default(0);
             $table->integer('twelve_balance')->default(0);
         });
